Refactor MineBitcoinBlockTenureChangeAndWaitForNakamotoBlock into MineBitcoinBlockAndTenureChange - and reuse the WaitForNakamotoBlock command
