+++ testnet/stacks-node/src/tests/signer/v0.rs
@@ -10965,8 +10965,8 @@ fn disallow_reorg_within_first_proposal_burn_block_timing_secs_but_more_than_one
         (MinerCommitOp::disable_for(test_context.clone(), MINER2)),
         BootToEpoch3,
         (MinerCommitOp::disable_for(test_context.clone(), MINER1)),
-        (MineBitcoinBlockTenureChangeAndWaitForNakamotoBlock::new(test_context.clone(), MINER1)),
-        //(WaitForNakamotoBlock::wait_from_global_height(test_context.clone(), MINER1)),
+        (MineBitcoinBlockAndTenureChange::new(test_context.clone(), MINER1)),
+        (WaitForNakamotoBlock::wait_from_state_height(test_context.clone(), MINER1)),
         (VerifyMinerWonSortition::new(test_context.clone(), MINER1)),
         (SubmitBlockCommit::new(test_context.clone(), MINER2)),
         (StacksMining::pause()),
